#Получение допуска. Исправляет отображаемое имя файла
