+++ add-project.php
@@ -1,6 +1,5 @@
 <?php
 require_once ('./config.php');
-require_once ('./src/constants.php');
 require_once ('./src/functions.php');
 require_once ('./src/db_connect.php');
 require_once ('./src/db_queries.php');
