+++ src/db_connect.php
@@ -11,7 +11,6 @@ if (!$connect) {
     die('Ошибка подключения к базе данных');
 }
 
-$show_complete_tasks = 0;
 $user_id = 0;
